test(vertex-ai): split embeddings tests
